feat: Add mike as a version provider
In order to support deviations and differences between multiple Cleura
Cloud regions, add version support via mike.

Define "main" as the default version, and update the build framework
so that we can treat each branch as a separate version.

Mike also requires that the user.name and user.email Git config
variables are set, in order for it to create its commit to
gh-pages. We can either set this with the "git config", or by using
the GIT_COMMITTER_NAME and GIT_COMMITTER_EMAIL environment variables
as defined by the CI configuration. The latter is the more flexible
option, so add GIT_* to the passenv list for the "mike" testenv.

References:
https://squidfunk.github.io/mkdocs-material/setup/setting-up-versioning/
https://github.com/jimporter/mike